% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllelFrequencyFoldChange.R
\name{AllelFrequencyFoldChange}
\alias{AllelFrequencyFoldChange}
\title{AllelFrequencyFoldChange}
\usage{
AllelFrequencyFoldChange(
  SE,
  VOI = NULL,
  group_of_interest,
  group1 = "group1",
  group2 = "group2",
  verbose = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{VOI}{The variants variants to be analyzed. If NULL all are used.}

\item{group_of_interest}{The column data that divides the cells.}

\item{group1}{The first group.}

\item{group2}{The second group.}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
Check if a cell is supported by a set of variants.
}
