% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSupplementing_big.R
\name{AmpliconSupplementing_big}
\alias{AmpliconSupplementing_big}
\title{Supplementing scRNAseq values with Amplicon values using big.matrix}
\usage{
AmpliconSupplementing_big(scRNAseq, amplicon)
}
\arguments{
\item{scRNAseq}{The SummarizedExperiment object containing the scRNAseq data.}

\item{amplicon}{The SummarizedExperiment object containing the amplicon data.}
}
\description{
We replace the values from an scRNAseq experiment with values we have from an amplicon experiment.
}
