% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAlleleFrequency.R
\name{CalculateAlleleFrequency}
\alias{CalculateAlleleFrequency}
\title{Calculating the Minor Allele Frequency.}
\usage{
CalculateAlleleFrequency(reference_reads, alternative_reads)
}
\arguments{
\item{reference_reads}{Reference reads matrix.}

\item{alternative_reads}{List of matrices for the alternative reads.}
}
\description{
We calculate the MAF for the MAEGATK results.
}
