% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAltReads.R
\name{CalculateAltReads}
\alias{CalculateAltReads}
\title{CalculateAltReads}
\usage{
CalculateAltReads(SE, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{chromosome_prefix}{List of matrices for the alternative reads.}
}
\description{
We calculate the number of reads covering a variant using forward and reverse reads.
}
