% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateConsensus.R
\name{CalculateConsensus}
\alias{CalculateConsensus}
\title{We calculate the consensus information from the MAEGATK results.}
\usage{
CalculateConsensus(SE, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{chromosome_prefix}{The chromosome name used as a prefix.}
}
\description{
We calculate the consensus information from the MAEGATK results.
}
