% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCorrelationPValue.R
\name{CalculateCorrelationPValue}
\alias{CalculateCorrelationPValue}
\title{We perform the correlation of SNVs and calculate the P values.}
\usage{
CalculateCorrelationPValue(
  variant_values,
  other_mutation,
  all_variants_list,
  min_intersecting_cells = 5
)
}
\arguments{
\item{variant_values}{other_mutation all_variants_list min_intersecting_cells}
}
\description{
We perform the correlation of SNVs and calculate the P values.
}
