% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCoverage.R
\name{CalculateCoverage}
\alias{CalculateCoverage}
\title{We calculate the consensus information from the MAEGATK results.}
\usage{
CalculateCoverage(alt_reads, ref_reads, chromosome_prefix)
}
\arguments{
\item{chromosome_prefix}{List of matrices for the alternative reads.}

\item{SE}{SummarizedExperiment object.}
}
\description{
We calculate the consensus information from the MAEGATK results.
}
