% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateForwardReads.R
\name{CalculateForwardReads}
\alias{CalculateForwardReads}
\title{CalculateForwardReads}
\usage{
CalculateForwardReads(SE, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{chromosome_prefix}{List of matrices for the alternative reads.}
}
\description{
We calculate the number of forward reads covering a variant.
}
