% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRefReads.R
\name{CalculateRefReads}
\alias{CalculateRefReads}
\title{CalculateRefReads}
\usage{
CalculateRefReads(SE, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{chromosome_prefix}{List of matrices for the reference reads.}
}
\description{
We calculate the number of reference reads covering a variant using forward and reverse reads.
}
