% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateReverseReads.R
\name{CalculateReverseReads}
\alias{CalculateReverseReads}
\title{CalculateReverseReads}
\usage{
CalculateReverseReads(SE, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{chromosome_prefix}{List of matrices for the alternative reads.}
}
\description{
We calculate the number of reverse reads covering a variant.
}
