% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateStrandCorrelation.R
\name{CalculateStrandCorrelation}
\alias{CalculateStrandCorrelation}
\title{CalculateStrandCorrelation}
\usage{
CalculateStrandCorrelation(SE, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{chromosome_prefix}{List of matrices for the alternative reads.}
}
\description{
We calculate the correlation between the amount of forward and reverse reads per variant.
}
