% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_to_numeric.R
\name{char_to_numeric}
\alias{char_to_numeric}
\title{char_to_numeric}
\usage{
char_to_numeric(char_value)
}
\arguments{
\item{char_value}{}
}
\description{
A function to convert the heterozyguous/homozyguous information from the VCF to the consensus information from VarTrix.
It is only used in LoadingVCF_typewise.R.
}
