% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalDefinition_Singular.R
\name{ClonalDefinition_Singular}
\alias{ClonalDefinition_Singular}
\title{Defining a single clone using multiple variants.}
\usage{
ClonalDefinition_Singular(
  se,
  variants,
  dropout = FALSE,
  variants_essential = NULL,
  variants_excluding = NULL,
  min_vaf = 0.01,
  clone_name = "SingularClone",
  verbose = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{variants}{Set of variants for clonal definition.}

\item{dropout}{The number of missing variants allowed.}

\item{variants_essential}{Which variants are considered essential? These will not be dropped if dropout is not NULL. The values can either be the variants names as character or indices for variants. A list of the same length as variants.}

\item{variants_excluding}{Which variants are not allowed in the clone? Variants names.  A list of the same length as variants.}

\item{min_vaf}{Minimum VAF for a cell to be considered positive.}

\item{clone_name}{What is the name of the clone? Default: SingularClone}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
This function defines a single clone using a set of variants. A cell is considered to be part of the clone, if it has all of the provided variants. A maximum number of missing variants can be set.
}
