% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalDiversity.R
\name{ClonalDiversity}
\alias{ClonalDiversity}
\title{Determining the clonal diversity.}
\usage{
ClonalDiversity(
  se,
  grouping = "Clones",
  cells = NULL,
  diversity_measure = "EffectiveSpecies",
  base = exp(1),
  verbose = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{grouping}{The meta data    column used.}

\item{cells}{Which cells should be used? NULL uses all cells.}

\item{diversity_measure}{What diversity measure should be calculated? Can be EffectiveSpecies, shannon, simpson, invsimpson}

\item{base}{The base for the diversity}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
This function determines the clonal diversity. The clones are defined using the function ClonalDefinition. It is not required to use clones, but any column in the column data of the object can be used.
}
