% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_NAMES.R
\name{combine_NAMES}
\alias{combine_NAMES}
\title{combine_NAMES}
\usage{
combine_NAMES(x, y)
}
\arguments{
\item{x}{First vector of names.}

\item{y}{Second vector of names.}
}
\description{
We combine two vectors of names.
}
