% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_SparseMatrix.R
\name{combine_SparseMatrix}
\alias{combine_SparseMatrix}
\title{combine_sparseMatrix}
\usage{
combine_SparseMatrix(matrix_1, matrix_2)
}
\arguments{
\item{matrix_1}{Your first sparse matrix.}

\item{matrix_2}{Your second matrix.}
}
\description{
We combine two sparse matrices
}
