% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAFMutMatrix.R
\name{computeAFMutMatrix}
\alias{computeAFMutMatrix}
\title{computeAFMutMatrix}
\usage{
computeAFMutMatrix(SE, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}
}
\description{
Calculate the allele frequency per variant.
Source: https://github.com/petervangalen/MAESTER-2021
}
