% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_Fisher.R
\name{Enrichment_FisherTest}
\alias{Enrichment_FisherTest}
\title{Enrichment_FisherTest}
\usage{
Enrichment_FisherTest(
  se = NULL,
  variant = NULL,
  use_nocall = FALSE,
  both_as_alt = TRUE,
  trait = NULL,
  trait_levels = NULL
)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{variant}{The variant to be analysed. The values for this variant are retrieved from the consensus assay.}

\item{use_nocall}{Should the NoCall values be used? Then the non genotyped cells would be a separate group in addition to Alt and Ref. If FALSE, the NoCalls are removed. Default = FALSE}

\item{both_as_alt}{Should the consensus value for heterozyguous (3 or both) be treated as homozyguous (Alt)? Default = TRUE}

\item{trait}{The trait the variant should be associated to. This can be any column from the colData of the object.}

\item{trait_levels}{Should only trait levels be used for the comparison? For example, if Sample 1-3 are present but only 1 and 2 should be used in the comparison use trait_levels = c("Sample_1", "Sample_2"). NULL uses all levels. Default = NULL}
}
\description{
This function performs the Fisher test. The specified variant is
}
