% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{Filtering}
\alias{Filtering}
\title{We load the matrix generated by LoadingVarTrix.R and filter the cells and genes.
We do this for one sample at a time.
We want to remove:
- all cells that are blacklisted,
- all cells that are not in the Seurat object,
- all cells that do not have at least one variant with >1 (Reference),
- all variants that are for alternative transcripts,
- all variants that are always NoCall.}
\usage{
Filtering(
  se,
  patient,
  blacklisted_barcodes_path = "NULL",
  fraction_threshold = 0.5,
  path_seurat = "NULL"
)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{patient}{Patient you are analysing. Character.}

\item{blacklisted_barcodes_path}{Barcodes you want to remove. Path to a file with one column without header.}

\item{fraction_threshold}{Variants with an VAF below this threshold are set to 0. Numeric.}

\item{path_seurat}{Path to a Seurat object.}
}
\description{
We load the matrix generated by LoadingVarTrix.R and filter the cells and genes.
We do this for one sample at a time.
We want to remove:
- all cells that are blacklisted,
- all cells that are not in the Seurat object,
- all cells that do not have at least one variant with >1 (Reference),
- all variants that are for alternative transcripts,
- all variants that are always NoCall.
}
