% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering_VAF_Threshold.R
\name{Filtering_VAF_Threshold}
\alias{Filtering_VAF_Threshold}
\title{Filtering_VAF_Threshold}
\usage{
Filtering_VAF_Threshold(SE, variant = NULL, thresholds = NULL, verbose = TRUE)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{variant}{Which variant should be evaluated? Cannot be more than 1 variant at a time.}

\item{thresholds}{The thresholds to be evaluated. Default between 0.01 and 0.99 in steps of 0.01.}

\item{verbose}{Should the function be verbose?}
}
\description{
We determine the number of mutated cells for a given variant using a vector of VAF thresholds.
}
