% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consensus.R
\name{get_consensus}
\alias{get_consensus}
\title{We get the consensus information for a specific matrix.}
\usage{
get_consensus(alt_base, ref_base, input_matrix, chromosome_prefix = "chrM")
}
\arguments{
\item{ref_base}{The reference base.}

\item{input_matrix}{Input matrix with the present reads numerically encoded.}

\item{chromosome_prefix}{The chromosome name used as a prefix.}

\item{letter}{The alternative base.}
}
\description{
We get the consensus information for a specific matrix.
}
