% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutMatrix.R
\name{getMutMatrix}
\alias{getMutMatrix}
\title{getMutMatrix}
\usage{
getMutMatrix(SE, cov, letter, ref_allele, chromosome_prefix)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{cov}{The coverage matrix from MAEGATK/MGATK.}

\item{letter}{The base we are interested in.}

\item{ref_allele}{Vector of reference alleles.}

\item{chromosome_prefix}{The chromosome prefix used.}
}
\description{
This function gets the allele frequency for a specific allele. It is used in computeAFMutMatrix.
Source: https://github.com/petervangalen/MAESTER-2021
}
