% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadMatrix.R
\name{getReadMatrix}
\alias{getReadMatrix}
\title{Get the counts for a specific base over all positions.}
\usage{
getReadMatrix(SE, letter, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{letter}{The base for which we want the counts.}

\item{chromosome_prefix}{The chromosome name used as a prefix.}
}
\description{
Get the counts for a specific base over all positions.
}
