% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefMatrix.R
\name{getRefMatrix}
\alias{getRefMatrix}
\title{We get the reference values from the MAEGATK result.}
\usage{
getRefMatrix(SE_object, letter, ref_alleles)
}
\arguments{
\item{SE_object}{letter ref_alleles}
}
\description{
We get the reference values from the MAEGATK result.
}
