% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVariantInfo.R
\name{GetVariantInfo}
\alias{GetVariantInfo}
\title{GetVariantInfo}
\usage{
GetVariantInfo(SE, information = "consensus", variants = NULL, cells = NULL)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{information}{The assay with the desired information. Default: consensus}

\item{variants}{A vector of variants.}

\item{cells}{A vector of cell IDs. On default all cells are returned. Default: NULL}
}
\description{
We get the genotyping information for a set of variants.
The function returns a matrix with the values from the specified assay.
}
