% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsci_pal.R
\name{ggsci_pal}
\alias{ggsci_pal}
\title{ggsci_pal}
\usage{
ggsci_pal(option, ...)
}
\arguments{
\item{option}{Your colour palette of choice.}

\item{...}{Further options passed to the palette function.}
}
\description{
Function to return colours from a ggsci palette.
}
\details{
The function returns a colour palette from ggsci.
Options are:
aaas: 10 d3: 10 futurama: 12 gsea: 12
igv: 51 jama: 7 jco: 10 npg: 10 lancet: 9 
locuszoom: 7 material: 10 nejm: 8 
rickandmorty: 12 simpsons: 16 
startrek: 7 tron: 7 uchicago: 9 
ucscgb: 26
}
