% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapVOI.R
\name{HeatmapVoi}
\alias{HeatmapVoi}
\title{HeatmapVoi}
\usage{
HeatmapVoi(
  SE,
  voi,
  annotation_trait = NULL,
  column_title = NULL,
  minimum_coverage = 0,
  remove_empty_cells = FALSE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{voi}{Variants Of Interest.}

\item{annotation_trait}{Cell Annotation at the bottom of the heat map.}

\item{column_title}{The title of the heat map. Default = NULL}

\item{minimum_coverage}{The minimum coverage per cell to be plotted.}

\item{remove_empty_cells}{Should cells that have a fraction of 0 for all variants be removed? Default = FALSE}
}
\description{
We plot a heatmap of a set of Variants Of Interest using the Variant Allele Frequency values of a SummarizedExperiment object.
}
