% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingMAEGATK.R
\name{LoadingMAEGATK}
\alias{LoadingMAEGATK}
\title{We load the MAEGATK output and transform it to be compatible with the VarTrix output.}
\usage{
LoadingMAEGATK(samples_path, vcf_path, vcf_path_MT, patient)
}
\arguments{
\item{samples_path}{Path to the csv file with the samples to be loaded.}

\item{vcf_path}{Path to the VCF file with the somatic variants.}

\item{vcf_path_MT}{Path to the VCF file with the MT variants.}

\item{patient}{The patient you want to load.}
}
\description{
We load the MAEGATK output and transform it to be compatible with the VarTrix output.
}
