% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingRawMatrix_typewise.R
\name{LoadingRawMatrix_typewise}
\alias{LoadingRawMatrix_typewise}
\title{LoadingRawMatrix_typewise}
\usage{
LoadingRawMatrix_typewise(
  samples_file,
  patient,
  patient_column = "patient",
  variant_use = NULL,
  cells_include = NULL,
  cells_exclude = NULL,
  matrix_column_separator = ",",
  matrix_header = TRUE,
  cellbarcode_index = 1,
  ref_index = 2,
  alt_index = 3,
  cellbarcode_length = 18,
  verbose = TRUE
)
}
\arguments{
\item{samples_file}{Path to the csv file with the samples to be loaded.}

\item{patient}{The patient you want to load.}

\item{patient_column}{The column that contains the patient information.}

\item{variant_use}{The variant of the respective matrices.}

\item{cells_include}{A vector of cell barcodes. Only these cells will be retained.}

\item{cells_exclude}{A vector of cell barcodes. These cells will be removed from the output.}

\item{cellbarcode_length}{The length of the cell barcode. This should be the length of the actual barcode plus two for the suffix (-1). Default = 18}

\item{verbose}{Should the function be verbose? Default = TRUE}

\item{samples_path}{Path to the input folder. Must include a barcodes file.}
}
\description{
A function to load a dense matrix of genotyping information and reformat it. Each matrix should have a column with the cell barcodes, one with the reference reads and one with the alternative reads.
The type in the design matrix is used as the variant name.
}
