% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingVarTrix_typewise.R
\name{LoadingVarTrix_typewise}
\alias{LoadingVarTrix_typewise}
\title{When we load all the different types of results (scRNAseq/amplicon and MT/amplicon),
we might need extreme amounts of memory. To solve this issue, I will load each type separately.
In a following function (AmpliconSupplementing), we can add the amplicon information to the
scRNAseq information.}
\usage{
LoadingVarTrix_typewise(
  samples_path,
  vcf_path,
  patient,
  type_use = "scRNAseq_Somatic"
)
}
\arguments{
\item{samples_path}{vcf_path patient type_use}
}
\description{
When we load all the different types of results (scRNAseq/amplicon and MT/amplicon),
we might need extreme amounts of memory. To solve this issue, I will load each type separately.
In a following function (AmpliconSupplementing), we can add the amplicon information to the
scRNAseq information.
}
