% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingVarTrix_typewise_big.R
\name{LoadingVarTrix_typewise_big}
\alias{LoadingVarTrix_typewise_big}
\title{Loading VarTrix results using big.matrix}
\usage{
LoadingVarTrix_typewise_big(
  samples_file,
  samples_path = NULL,
  vcf_path,
  patient,
  type_use = "scRNAseq_Somatic",
  min_reads = 3
)
}
\arguments{
\item{samples_file}{Path to the csv file with the samples to be loaded.}

\item{samples_path}{Path to the input folder. Must include a barcodes file.}

\item{vcf_path}{Path to the VCF file with the variants.}

\item{patient}{The patient you want to load.}

\item{type_use}{The type of input. Has to be one of: scRNAseq_Somatic, Amplicon_Somatic, scRNAseq_MT, Amplicon_MT.}

\item{min_reads}{The minimum number of reads we want. Otherwise we treat this as a NoCall.}
}
\description{
When we load all the different types of results (scRNAseq/amplicon and MT/amplicon),
we might need extreme amounts of memory. To solve this issue, I will load each type separately.
In a following function (AmpliconSupplementing), we can add the amplicon information to the
scRNAseq information.
}
