% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingVCF_typewise.R
\name{LoadingVCF_typewise}
\alias{LoadingVCF_typewise}
\title{LoadingVCF_typewise}
\usage{
LoadingVCF_typewise(
  samples_file,
  samples_path = NULL,
  barcodes_path = NULL,
  vcf_path,
  patient,
  sample = NULL,
  type_use = "scRNAseq_Somatic",
  min_reads = NULL,
  min_cells = 2
)
}
\arguments{
\item{samples_file}{Path to the csv file with the samples to be loaded.}

\item{samples_path}{Path to the input folder. Must include a barcodes file.}

\item{vcf_path}{Path to the VCF file with the variants.}

\item{patient}{The patient you want to load.}

\item{type_use}{The type of input. Has to be one of: scRNAseq_Somatic, Amplicon_Somatic, scRNAseq_MT, Amplicon_MT.}

\item{min_reads}{The minimum number of reads we want. Otherwise we treat this as a NoCall. Default = NULL.}

\item{min_cells}{The minimum number of cells for a variant. Otherwise, we will remove a variant. Default = 2.}
}
\description{
We load a cellwise pileup result from a VCF file.
}
