% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merging_SE_list.R
\name{Merging_SE_list}
\alias{Merging_SE_list}
\title{Merging list of SummarizedExperiment objects.}
\usage{
Merging_SE_list(se)
}
\arguments{
\item{se}{SummarizedExperiment object}
}
\description{
This function is a wrapper for do.all("cbind", se).
}
