% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowWiseSplit.R
\name{RowWiseSplit}
\alias{RowWiseSplit}
\title{We get the filtered results, we have from the 01_FilteringSampleWise. We now split the results into the
individual columns and save them in a list object.}
\usage{
RowWiseSplit(se, n_cores = 1)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{n_cores}{Number of cores to use.}
}
\description{
We get the filtered results, we have from the 01_FilteringSampleWise. We now split the results into the
individual columns and save them in a list object.
}
