% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_object.R
\name{save_object}
\alias{save_object}
\title{We define a loading function to save the RDS files quicker.}
\usage{
save_object(object, file_name, file_format = NULL)
}
\arguments{
\item{object}{file_name file_format}
}
\description{
We define a loading function to save the RDS files quicker.
}
