% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdiv.R
\name{sdiv}
\alias{sdiv}
\title{Division of sparse matrix.}
\usage{
sdiv(X, Y, names = dimnames(X))
}
\arguments{
\item{X}{First sparse matrix.}

\item{Y}{Second sparse matrix.}

\item{names}{The dimension names (dimnames(X)).}
}
\description{
Division of sparse matrix.
}
