% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparatingMatrixToList.R
\name{SeparatingMatrixToList}
\alias{SeparatingMatrixToList}
\title{We separate a matrix of variant information to a list.
Each variant is an entry in the list.}
\usage{
SeparatingMatrixToList(row_to_save, total_matrix)
}
\arguments{
\item{row_to_save}{total_matrix}
}
\description{
We separate a matrix of variant information to a list.
Each variant is an entry in the list.
}
