% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMI_seq.R
\name{UMI_seq}
\alias{UMI_seq}
\title{UMI_seq}
\usage{
UMI_seq(number, umi_length = 12)
}
\arguments{
\item{number}{The numeric expression of a UMI.}

\item{umi_length}{The length of the UMI. Default = 12}
}
\description{
This function converts the numeric UMI representation found in a 10X molecule_info.h5 file to a character string.
}
