% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantBurden.R
\name{VariantBurden}
\alias{VariantBurden}
\title{VariantBurden}
\usage{
VariantBurden(se)
}
\arguments{
\item{se}{SummarizedExperiment object}
}
\description{
Calculate the variant burden per cell.
We simply sum up the MAF values per cell.
}
