% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantBurden_big.R
\name{VariantBurden_big}
\alias{VariantBurden_big}
\title{Calculate the variant burden per cell using big matrices.}
\usage{
VariantBurden_big(se)
}
\arguments{
\item{se}{SummarizedExperiment object}
}
\description{
Calculate the variant burden per cell using big matrices.
}
