% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantCloneSizeThresholding.R
\name{VariantCloneSizeThresholding}
\alias{VariantCloneSizeThresholding}
\title{We get variants of interest using a clone size thresholding.}
\usage{
VariantCloneSizeThresholding(
  se,
  min_coverage = 2,
  fraction_negative_cells = 0.9,
  min_clone_size = 10,
  vaf_threshold = 0.5
)
}
\arguments{
\item{se}{min_coverage fraction_negative_cells min_clone_size vaf_threshold}
}
\description{
We get variants of interest using a clone size thresholding.
}
