% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantCorrelationHeatmap.R
\name{VariantCorrelationHeatmap}
\alias{VariantCorrelationHeatmap}
\title{We generate a heatmap showing the correlation of somatic variants with the MT variants.}
\usage{
VariantCorrelationHeatmap(
  correlation_results,
  output_path,
  patient,
  min_alt_cells = 5,
  min_correlation = 0.5,
  width_use = 2000,
  height_use = 2000,
  padding_use = c(165, 165, 2, 2)
)
}
\arguments{
\item{correlation_results}{output_path patient min_alt_cellsmin_correlation width_use height_use padding_use}
}
\description{
We generate a heatmap showing the correlation of somatic variants with the MT variants.
}
