% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantFisherTestHeatmap.R
\name{VariantFisherTestHeatmap}
\alias{VariantFisherTestHeatmap}
\title{VariantFisherTestHeatmap}
\usage{
VariantFisherTestHeatmap(
  fisher_results,
  patient,
  min_alt_cells = 5,
  min_oddsratio = 1
)
}
\arguments{
\item{fisher_results}{Data.frame with the correlation results.}

\item{patient}{The patient for this heatmap.}

\item{min_alt_cells}{Minimum number of mutated cells needed, otherwise an association will not be plotted.}

\item{min_oddsratio}{Minimum correlation needed.}
}
\description{
We generate a heatmap showing the Fisher test of somatic variants with the MT variants.
}
