% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantQuantileThresholding_Combined.R
\name{VariantQuantileThresholding_Combined}
\alias{VariantQuantileThresholding_Combined}
\title{VariantQuantileThresholding_Combined}
\usage{
VariantQuantileThresholding_Combined(
  SE,
  min_coverage = 2,
  quantiles = c(0.1, 0.9),
  thresholds = c(0.1, 0.9),
  top_cells = NULL,
  top_VAF = NULL,
  min_quality = NULL,
  mean_allele_frequency = 0,
  group_of_interest = NULL,
  group1 = NULL,
  group2 = NULL,
  group_factor = NULL,
  remove_nocall = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{min_coverage}{Minimum coverage needed.}

\item{quantiles}{The lower and upper quantile you want to use.}

\item{thresholds}{The VAF thresholds you want to use for the quantiles.}

\item{top_cells}{The number of cells with at least top_VAF percent for a variant.}

\item{top_VAF}{The VAF for the top cells.}

\item{min_quality}{The minimum quality you want for the Variants of Interest. Can be ignored by setting it to NULL.}

\item{mean_allele_frequency}{The minimum mean allele frequency. Default = 0}

\item{group_of_interest}{The column data that divides the cells.}

\item{group1}{The first group of interest. If set, the quantiles are only calculated for this group.}

\item{group2}{The second group of interest.}

\item{group_factor}{How much higher has the mean allele frequency to be in group 1 when compared to group 2?}

\item{remove_nocall}{Should NoCall cells (consensus = 0) be disregarded during the analysis?}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
We get variants of interest using the quantile thresholding. This function combines the functions VariantSelection_Quantile, VariantSelection_Group and VariantSelection_TopCells.
If you use top_cells and top_VAF, you have to only supply one quantil value (quantiles = 0.9, thresholds = 0).
This function is adapted from the Peter van Galen.
Source: https://github.com/petervangalen/MAESTER-2021
}
