% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSelection_Quantile.R
\name{VariantSelection_Quantile}
\alias{VariantSelection_Quantile}
\title{VariantSelection_Quantile}
\usage{
VariantSelection_Quantile(
  SE,
  min_coverage = 2,
  quantiles = c(0.1, 0.9),
  thresholds = c(0.1, 0.9),
  min_quality = NULL,
  mean_allele_frequency = 0,
  remove_nocall = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{min_coverage}{Minimum coverage needed.}

\item{quantiles}{The lower and upper quantile you want to use.}

\item{thresholds}{The VAF thresholds you want to use for the quantiles.}

\item{min_quality}{The minimum quality you want for the Variants of Interest. Can be ignored by setting it to NULL.}

\item{mean_allele_frequency}{The minimum mean allele frequency. Default = 0}

\item{remove_nocall}{Should NoCall cells (consensus = 0) be disregarded during the analysis?}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
We get variants of interest using the quantile thresholding.
This function is adapted from the Peter van Galen.
Source: https://github.com/petervangalen/MAESTER-2021
}
