% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSelection_TopCells.R
\name{VariantSelection_TopCells}
\alias{VariantSelection_TopCells}
\title{VariantSelection_TopCells}
\usage{
VariantSelection_TopCells(
  SE,
  min_coverage = 2,
  quantiles = 0.9,
  thresholds = 0,
  top_cells = 10,
  top_VAF = 0.5,
  min_quality = NULL,
  remove_nocall = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{min_coverage}{Minimum coverage needed.}

\item{quantiles}{The lower and upper quantile you want to use.}

\item{thresholds}{The VAF thresholds you want to use for the quantiles.}

\item{min_quality}{The minimum quality you want for the Variants of Interest. Can be ignored by setting it to NULL.}

\item{remove_nocall}{Should NoCall cells (consensus = 0) be disregarded during the analysis?}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
We get variants of interest by selecting a small number of cells with a high VAF for a variant.
This function is adapted from the Peter van Galen.
Source: https://github.com/petervangalen/MAESTER-2021
}
