% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSelection_VMR.R
\name{VariantSelection_VMR}
\alias{VariantSelection_VMR}
\title{VariantSelection_VMR}
\usage{
VariantSelection_VMR(
  SE,
  stabilize_variance = TRUE,
  low_coverage_threshold = 10,
  minimum_fw_rev_reads = 2,
  verbose = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{stabilize_variance}{Should the variance be stabilized by using the mean allele frequency for cells with a low coverage? Coverage threshold is set by low_coverage_threshold.}

\item{low_coverage_threshold}{Cells below this threshold are set to the mean.}

\item{minimum_fw_rev_reads}{How many forward and reverse reads should a cell have? Default = 2}

\item{verbose}{Should the function be verbose?}
}
\description{
This is an adaption of the VMR based selection of variants from Miller et al. 
This selection process is designed for ATAC data and not scRNAseq data.
}
