% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSelection_VMR_Group.R
\name{VariantSelection_VMR_Group}
\alias{VariantSelection_VMR_Group}
\title{VariantSelection_VMR_Group}
\usage{
VariantSelection_VMR_Group(
  SE,
  stabilize_variance = TRUE,
  low_coverage_threshold = 10,
  minimum_fw_rev_reads = 2,
  group_of_interest = NULL,
  group1 = NULL,
  group2 = NULL,
  group_factor = 5,
  verbose = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{stabilize_variance}{Should the variance be stabilized by using the mean allele frequency for cells with a low coverage? Coverage threshold is set by low_coverage_threshold.}

\item{low_coverage_threshold}{Cells below this threshold are set to the mean.}

\item{minimum_fw_rev_reads}{How many forward and reverse reads should a cell have? Default = 2}

\item{group_of_interest}{The group of interest in the column data.}

\item{group1}{The group of interest.}

\item{group2}{The second group.}

\item{group_factor}{How much higher should the VAF be in group 1 comapred to group 2? Default = 5}

\item{verbose}{Should the function be verbose?}
}
\description{
This is an adaption of the VMR based selection of variants from Miller et al. 
This selection process is designed for ATAC data and not scRNAseq data. 
The variants are first selected as described. Then the average allele frequency between the the two groups of interest are compared.
If the allele frequency in group1 is higher than the required threshold, the variant is retained.
}
