% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantWiseFisherTest.R,
%   R/VariantWiseFisherTest2.R
\name{VariantWiseFisherTest}
\alias{VariantWiseFisherTest}
\title{We perfrom the Fisher test to determine which variants are associated.}
\usage{
VariantWiseFisherTest(variants_list, n_cores = 1, p_value_adjustment = "fdr")

VariantWiseFisherTest(variants_list, n_cores = 1, p_value_adjustment = "fdr")
}
\arguments{
\item{variants_list}{List of fraction values.}

\item{n_cores}{Number of cores you want to use. Numeric.}

\item{p_value_adjustment}{Method for P value adjustment. See p.adjust for details.}
}
\description{
We perfrom the Fisher test to determine which variants are associated.

We perfrom the Fisher test to determine which variants are associated.
}
