% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllelFrequencyFoldChange.R
\name{AllelFrequencyFoldChange}
\alias{AllelFrequencyFoldChange}
\title{AllelFrequencyFoldChange}
\usage{
AllelFrequencyFoldChange(
  SE,
  VOI = NULL,
  group_of_interest,
  group1 = "group1",
  group2 = "group2",
  maximum_foldchange = NULL,
  minimum_foldchange = NULL,
  minimum_coverage = NULL,
  minimum_allele_freq = NULL,
  maximum_allele_freq = NULL,
  verbose = FALSE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{VOI}{The variants variants to be analyzed. If NULL all are used.}

\item{group_of_interest}{The column data that divides the cells.}

\item{group1}{The first group.}

\item{group2}{The second group.}

\item{maximum_foldchange}{Maximum fold change.}

\item{minimum_foldchange}{Minimum fold change.}

\item{minimum_coverage}{Minimum coverage for a variant.}

\item{minimum_allele_freq}{Minimum allele frequency in both groups.}

\item{maximum_allele_freq}{Maximum allele frequency in both groups.}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
Check if a cell is supported by a set of variants.
}
