% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSupplementing.R
\name{AmpliconSupplementing}
\alias{AmpliconSupplementing}
\title{Supplementing scRNAseq values with Amplicon values}
\usage{
AmpliconSupplementing(scRNAseq, amplicon)
}
\arguments{
\item{scRNAseq}{amplicon}
}
\description{
Supplementing scRNAseq values with Amplicon values
}
