% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAlleleFrequency.R
\name{CalculateAlleleFrequency}
\alias{CalculateAlleleFrequency}
\title{Calculating the Minor Allele Frequency.}
\usage{
CalculateAlleleFrequency(reference_reads, alternative_reads, pseudo_count = 0)
}
\arguments{
\item{reference_reads}{Reference reads matrix.}

\item{alternative_reads}{List of matrices for the alternative reads.}

\item{pseudo_count}{= What is the pseudo count you want to add to the reference_reads matrix. Default = 0}
}
\description{
We calculate the MAF from a reference reads matrix and an alternative reads matrix.
This function is intended to be used with the mitochondrial genome and not with other somatic mutations.
}
