% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateConsensus.R
\name{CalculateConsensus}
\alias{CalculateConsensus}
\title{We calculate the consensus information from the MAEGATK results.}
\usage{
CalculateConsensus(input_se, input_base)
}
\arguments{
\item{input_se}{Input SummarizedExperiment object.}

\item{input_base}{The reference allele we want to use. Use input_base <- c("A", "C", "G", "T") with lapply.}
}
\description{
We calculate the consensus information from the MAEGATK results.
}
