% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateConsensus.R
\name{CalculateConsensus}
\alias{CalculateConsensus}
\title{We calculate the consensus information from the MAEGATK results.}
\usage{
CalculateConsensus(input_se, input_base)
}
\arguments{
\item{input_se}{input_base}
}
\description{
We calculate the consensus information from the MAEGATK results.
}
