% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateConsensus.R
\name{CalculateConsensus}
\alias{CalculateConsensus}
\title{CalculateConsensus}
\usage{
CalculateConsensus(SE, chromosome_prefix = "chrM", verbose = FALSE)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{chromosome_prefix}{The chromosome name used as a prefix.}

\item{verbose}{Should the function be verbose? Default = FALSE}
}
\description{
We calculate the consensus information from the MAEGATK results.
We set cells that have only alternative reads to 2 (Alternative).
We set cells that have only reference reads to 1 (Reference).
We set cells that have a mixture of alternative and reference reads to 3 (Both).
We set cells that have no reads to 0 (NoCall).

Please note. Cells can have reads for the reference of a specific variant and no reads for the alternative.
The cell can still have a reads for the other alternative alleles. Then the cell is still considered as 0 (NoCall) for this variant.
For example:
A cell has at position 3: 0 A reads, 53 T reads, 63 C reads, 148 T reads.
For the variant chrM_3_T_A, the cell would have 53 reference reads, but also reads for other variants at this position.
To make sure that there is no confusion, the cell is set to NoCall.
}
