% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCorrelationPValue.R
\name{CalculateCorrelationPValue}
\alias{CalculateCorrelationPValue}
\title{Correlating the SNVs}
\usage{
CalculateCorrelationPValue(
  variant_values,
  other_mutation,
  all_variants_list,
  value_type = "consensus",
  min_intersecting_cells = 5
)
}
\arguments{
\item{variant_values}{The fraction values you are analyzing. A vector.}

\item{other_mutation}{All other variants you have. A vector of variant names.}

\item{all_variants_list}{List of fraction values for all the variants you want to compare your variant with.}

\item{value_type}{Are we using consensus or other information?}

\item{min_intersecting_cells}{Minimum number of intersecting cells. Correlations with less than this will not be performed.}
}
\description{
We perform the correlation of SNVs and calculate the P values.
}
