% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCoverage.R
\name{CalculateCoverage}
\alias{CalculateCoverage}
\title{CalculateCoverage}
\usage{
CalculateCoverage(SE, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{chromosome_prefix}{List of matrices for the alternative reads.}
}
\description{
We calculate the coverage information per variant from the MAEGATK results.
}
