% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateQuality.R
\name{CalculateQuality}
\alias{CalculateQuality}
\title{CalculateQuality}
\usage{
CalculateQuality(SE, variants, chromosome_prefix = "chrM")
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{variants}{The variants you want to get the quality for.}

\item{chromosome_prefix}{List of matrices for the alternative reads.}
}
\description{
We calculate the quality per variant.
}
