% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineSEobjects.R
\name{CombineSEobjects}
\alias{CombineSEobjects}
\title{We combine two SummarizedExperiment objects.}
\usage{
CombineSEobjects(se_somatic, se_MT, suffixes = c("_somatic", "_MT"))
}
\arguments{
\item{se_somatic}{se_MT suffixes min_intersecting_cells}
}
\description{
We combine two SummarizedExperiment objects.
}
