% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineSEobjects.R
\name{CombineSEobjects}
\alias{CombineSEobjects}
\title{We combine two SummarizedExperiment objects.}
\usage{
CombineSEobjects(se_somatic, se_MT, suffixes = c("_somatic", "_MT"))
}
\arguments{
\item{se_somatic}{SummarizedExperiment object for the somatic variants.}

\item{se_MT}{SummarizedExperiment object for the MT variants.}

\item{suffixes}{The suffixes you want to add to the meta data.frame.}
}
\description{
We combine two SummarizedExperiment objects.
}
