% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAFMutMatrix.R
\name{computeAFMutMatrix}
\alias{computeAFMutMatrix}
\title{Calculate the allele frequency per variant.}
\usage{
computeAFMutMatrix(SE)
}
\arguments{
\item{SE}{SummarizedExperiment object.}
}
\description{
Calculate the allele frequency per variant.
}
