% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{Filtering}
\alias{Filtering}
\title{We load the matrix generated by LoadingVarTrix.R and filter the cells and genes.
We do this for one sample at a time.
We want to remove:
- all cells that are blacklisted,
- all cells that are not in the Seurat object,
- all cells that do not have at least one variant with >1 (Reference),
- all variants that are for alternative transcripts,
- all variants that are always NoCall.}
\usage{
Filtering(
  se,
  patient,
  blacklisted_barcodes_path = "NULL",
  fraction_threshold = 0.5,
  path_seurat = "NULL"
)
}
\arguments{
\item{se}{patient blacklisted_barcodes_path fraction_threshold path_seurat}
}
\description{
We load the matrix generated by LoadingVarTrix.R and filter the cells and genes.
We do this for one sample at a time.
We want to remove:
- all cells that are blacklisted,
- all cells that are not in the Seurat object,
- all cells that do not have at least one variant with >1 (Reference),
- all variants that are for alternative transcripts,
- all variants that are always NoCall.
}
