% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{Filtering}
\alias{Filtering}
\title{Filtering the loaded genotyping data.}
\usage{
Filtering(
  se,
  blacklisted_barcodes_path = NULL,
  fraction_threshold = NULL,
  alts_threshold = NULL,
  path_seurat = NULL,
  min_cells_per_variant = 2,
  min_variants_per_cell = 1
)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{blacklisted_barcodes_path}{Barcodes you want to remove. Path to a file with one column without header.}

\item{fraction_threshold}{Variants with an VAF below this threshold are set to 0. Numeric. Default = NULL.}

\item{alts_threshold}{Variants with a number of alt reads less than this threshold are set to 0. Numeric. Default = NULL.}

\item{path_seurat}{Path to a Seurat object. Cells not present in the object will be removed.}

\item{min_cells_per_variant}{In how many cells should a variant be present to be included? Numeric. Default = 2.}

\item{min_variants_per_cell}{How many variants should be covered in a cell have to be included? Default = 1.}
}
\description{
We filter a SummarizedExperiment object to exclude variants and cells.
}
\details{
We do this for one sample at a time.
We want to remove:
\enumerate{
  \item all cells that are blacklisted,
  \item all cells that are not in a Seurat object,
  \item all cells that do not have at least one variant with >1 (Reference),
  \item all variants that are for alternative transcripts,
  \item all variants that are always NoCall,
  \item set variants with a VAF below a threshold to reference.
}
}
\examples{
\dontrun{
  # Removing all variants that are not detected in at least 2 cells.
  # Before we remove the variants, we set fraction value of variants below 0.05 to 0.
  se_geno <- Filtering(se_geno, min_cells_per_variant = 2, fraction_threshold = 0.05)
}
}
