% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAltMatrix.R
\name{getAltMatrix}
\alias{getAltMatrix}
\title{We get the alt values from the MAEGATK results.}
\usage{
getAltMatrix(SE_object, letter, ref_alleles)
}
\arguments{
\item{SE_object}{SummarizedExperiment object.}

\item{letter}{The base you want to use. Character.}

\item{ref_alleles}{The reference alleles.}
}
\description{
We get the alt values from the MAEGATK results.
}
