% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellInfoPerVariant.R
\name{GetCellInfoPerVariant}
\alias{GetCellInfoPerVariant}
\title{We get the variant information per cell.}
\usage{
GetCellInfoPerVariant(se, voi_ch)
}
\arguments{
\item{se, }{voi_ch voi_ch}
}
\description{
We get the variant information per cell.
}
