% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellInfoPerVariant.R
\name{GetCellInfoPerVariant}
\alias{GetCellInfoPerVariant}
\title{We get the variant information per cell.}
\usage{
GetCellInfoPerVariant(se, voi_ch, verbose = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{voi_ch}{Variants of interest.}

\item{verbose}{Should the function be verbose? Default = FALSE}
}
\description{
We get the variant information per cell.
}
