% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefMatrix.R
\name{getRefMatrix}
\alias{getRefMatrix}
\title{We get the reference values from the MAEGATK result.}
\usage{
getRefMatrix(SE_object, letter, chromosome_prefix = "chrM")
}
\arguments{
\item{SE_object}{SummarizedExperiment object.}

\item{letter}{The base you are analysing. You get a matrix that shows which cells have how many referecen reads for this letter.}

\item{chromosome_prefix}{The chromosome prefix used.}

\item{ref_allele}{The reference alleles.}
}
\description{
We get the reference values from the MAEGATK result.
}
