% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefMatrix.R
\name{getRefMatrix}
\alias{getRefMatrix}
\title{getRefMatrix}
\usage{
getRefMatrix(SE_object, letter, chromosome_prefix = "chrM")
}
\arguments{
\item{SE_object}{SummarizedExperiment object.}

\item{letter}{The base you are analysing. You get a matrix that shows which cells have how many reference reads for this letter.}

\item{chromosome_prefix}{The chromosome prefix used.}
}
\description{
We get the reference values from the MAEGATK result.
Source: https://github.com/petervangalen/MAESTER-2021
}
