% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapVOI.R
\name{HeatmapVoi}
\alias{HeatmapVoi}
\title{HeatmapVoi}
\usage{
HeatmapVoi(SE, voi, annotation_trait = NULL)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{voi}{Variants Of Interest.}

\item{annotation_trait}{Cell Annotation at the bottom of the heat map.}
}
\description{
We plot a heatmap of a set of Variants Of Interest using the Variant Allele Frequency values of a SummarizedExperiment object.
}
