% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapVOI.R
\name{HeatmapVoi}
\alias{HeatmapVoi}
\title{HeatmapVoi}
\usage{
HeatmapVoi(
  SE,
  voi,
  annotation_trait = NULL,
  column_title = NULL,
  minimum_coverage = 0,
  sort_cells = FALSE,
  remove_empty_cells = FALSE,
  minimum_allele_freq = 0,
  cluster_variants = FALSE,
  cluster_variants_distance = "euclidean",
  cluster_variants_method = "complete"
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{voi}{Variants Of Interest.}

\item{annotation_trait}{Cell Annotation at the bottom of the heat map.}

\item{column_title}{The title of the heat map. Default = NULL}

\item{minimum_coverage}{The minimum coverage per cell to be plotted.}

\item{sort_cells}{Should the cells be sorted by ordering the cells according the the largest clones? FALSE uses default complete clustering with euclidean distance.}

\item{remove_empty_cells}{Should cells that have a fraction of 0 for all variants be removed? Default = FALSE}

\item{minimum_allele_freq}{Minimum allele frequency to include a cell.}

\item{cluster_variants}{Should the variants be clustered? Default FALSE.}

\item{cluster_variants_distance}{The distance for clustering the variants. Default euclidean.}

\item{cluster_variants_method}{The distance for the variant clustering. Default complete}
}
\description{
We plot a heatmap of a set of Variants Of Interest using the Variant Allele Frequency values of a SummarizedExperiment object.
}
