% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_object.R
\name{load_object}
\alias{load_object}
\title{We define a loading function to load the RDS files quicker.}
\usage{
load_object(file_name)
}
\arguments{
\item{file_name}{The path to the file.}
}
\description{
We define a loading function to load the RDS files quicker.
}
