% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_object.R
\name{load_object}
\alias{load_object}
\title{load_object}
\usage{
load_object(file_name)
}
\arguments{
\item{file_name}{The path to the file.}
}
\description{
A loading function to load the RDS files quicker.
Source: https://github.com/CostaLab/CimpleG
}
