% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingMAEGATK_typewise.R
\name{LoadingMAEGATK_typewise}
\alias{LoadingMAEGATK_typewise}
\title{LoadingMAEGATK_typewise}
\usage{
LoadingMAEGATK_typewise(
  samples_file,
  patient,
  samples_path = NULL,
  patient_column = "patient",
  type_use = "scRNAseq_MT",
  chromosome_prefix = "chrM",
  min_cells = 2,
  cells_include = NULL,
  cells_exclude = NULL,
  barcodes_path = NULL,
  cellbarcode_length = 18,
  verbose = TRUE
)
}
\arguments{
\item{samples_file}{Path to the csv file with the samples to be loaded.}

\item{patient}{The patient you want to load.}

\item{samples_path}{Path to the input folder.}

\item{patient_column}{The column that contains the patient information. Use merge, if all samples should be merged.}

\item{type_use}{The type of input. Only rows that have the specified type will be loaded.}

\item{chromosome_prefix}{The prefix you want use. Default: "chrM"}

\item{min_cells}{The minimum number of cells with coverage for a variant. Variants with coverage in less than this amount of cells are removed. Default = 2}

\item{cells_include}{A vector of cell barcodes. Only these cells will be retained.}

\item{cells_exclude}{A vector of cell barcodes. These cells will be removed from the output.}

\item{barcodes_path}{Path to the barcodes file tsv. Default = NULL}

\item{cellbarcode_length}{The length of the cell barcode. This should be the length of the actual barcode plus two for the suffix (-1). Default = 18}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
We load the MAEGATK output and transform it to be compatible with the VarTrix output.
The input file is a specifically formated csv file with all the necessary information to run the analysis.
Note that the source column in the input file needs to be mgaetk or mgatk for this function. This is case insensitive.
If you want to only load a single sample without the use of an input file, you have to set the following variables.
\enumerate{
  \item samples_path
  \item barcodes_path
  \item patient
  \item samples_file = NULL
}
}
