% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingVarTrix.R
\name{LoadingVarTrix}
\alias{LoadingVarTrix}
\title{We load the VarTrix results for one patient and merge all types together.}
\usage{
LoadingVarTrix(samples_path, vcf_path, vcf_path_MT, patient)
}
\arguments{
\item{samples_path}{vcf_path vcf_path_MT patient}
}
\description{
We load the VarTrix results for one patient and merge all types together.
}
