% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingVarTrix_typewise.R
\name{LoadingVarTrix_typewise}
\alias{LoadingVarTrix_typewise}
\title{When we load all the different types of results (scRNAseq/amplicon and MT/amplicon),
we might need extreme amounts of memory. To solve this issue, I will load each type separately.
In a following function (AmpliconSupplementing), we can add the amplicon information to the
scRNAseq information.}
\usage{
LoadingVarTrix_typewise(
  samples_path,
  vcf_path,
  patient,
  type_use = "scRNAseq_Somatic",
  chromosome_prefix = "chrM"
)
}
\arguments{
\item{samples_path}{Path to the csv file with the samples to be loaded.}

\item{vcf_path}{Path to the VCF file with the variants.}

\item{patient}{The patient you want to load.}

\item{type_use}{The type of input. Has to be one of: scRNAseq_Somatic, Amplicon_Somatic, scRNAseq_MT, Amplicon_MT.}
}
\description{
When we load all the different types of results (scRNAseq/amplicon and MT/amplicon),
we might need extreme amounts of memory. To solve this issue, I will load each type separately.
In a following function (AmpliconSupplementing), we can add the amplicon information to the
scRNAseq information.
}
