% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingVarTrix_typewise.R
\name{LoadingVarTrix_typewise}
\alias{LoadingVarTrix_typewise}
\title{When we load all the different types of results (scRNAseq/amplicon and MT/amplicon),
we might need extreme amounts of memory. To solve this issue, I will load each type separately.
In a following function (AmpliconSupplementing), we can add the amplicon information to the
scRNAseq information.}
\usage{
LoadingVarTrix_typewise(
  samples_file,
  samples_path = NULL,
  barcodes_path = NULL,
  snp_path = NULL,
  vcf_path,
  patient,
  sample = NULL,
  type_use = "scRNAseq_Somatic",
  min_reads = 3,
  min_cells = 2
)
}
\arguments{
\item{samples_file}{Path to the csv file with the samples to be loaded.}

\item{samples_path}{Path to the input folder. Must include a barcodes file.}

\item{vcf_path}{Path to the VCF file with the variants.}

\item{patient}{The patient you want to load.}

\item{type_use}{The type of input. Has to be one of: scRNAseq_Somatic, Amplicon_Somatic, scRNAseq_MT, Amplicon_MT.}

\item{min_reads}{The minimum number of reads we want. Otherwise we treat this as a NoCall. Not yet implemented.}

\item{min_cells}{The minimum number of cells for a variant. Otherwise, we will remove a variant.}
}
\description{
When we load all the different types of results (scRNAseq/amplicon and MT/amplicon),
we might need extreme amounts of memory. To solve this issue, I will load each type separately.
In a following function (AmpliconSupplementing), we can add the amplicon information to the
scRNAseq information.
}
