% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingVCF_typewise.R
\name{LoadingVCF_typewise}
\alias{LoadingVCF_typewise}
\title{LoadingVCF_typewise}
\usage{
LoadingVCF_typewise(
  samples_file,
  samples_path = NULL,
  barcodes_path = NULL,
  vcf_path,
  patient,
  type_use = "scRNAseq_Somatic",
  min_reads = NULL,
  min_cells = 2,
  remove_N_alternative = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{samples_file}{Path to the csv file with the samples to be loaded.}

\item{samples_path}{Path to the input folder. Must include a barcodes file.}

\item{barcodes_path}{Path to the cell barcodes tsv. Default = NULL}

\item{vcf_path}{Path to the VCF file with the variants.}

\item{patient}{The patient you want to load.}

\item{type_use}{The type of input. Has to be one of: scRNAseq_Somatic, Amplicon_Somatic, scRNAseq_MT, Amplicon_MT.}

\item{min_reads}{The minimum number of reads we want. Otherwise we treat this as a NoCall. Default = NULL.}

\item{min_cells}{The minimum number of cells for a variant. Otherwise, we will remove a variant. Default = 2.}

\item{remove_N_alternative}{Remove all variants that have N as an alternative, see Description. Default = TRUE}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
We load a cellwise pileup result from a VCF file.
If you want to only load a single sample without the use of an input file, you have to set the following variables.
\enumerate{
  \item samples_path
  \item barcodes_path
  \item patient
  \item samples_file = NULL
}

It has happened that reads with an N allele were aligned. This can cause problems since these variants are typically not in variants lists.
We can remove all of these variants by setting remove_N_alternative to TRUE (the default).
Set this option to FALSE, if you really want to retain these variants.
}
