% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowWiseSplit.R
\name{RowWiseSplit}
\alias{RowWiseSplit}
\title{RowWiseSplit}
\usage{
RowWiseSplit(se, n_cores = 1, minimum_reads, remove_nocalls = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{n_cores}{Number of cores to use.}

\item{remove_nocalls}{Do you want to remove NoCall cells?}
}
\description{
Performing the correlation or Fisher test association for a SummarizedExperiment object requires extreme amounts of memory.
To reduce the amount of memory necessary, we instead get the individual rows from the consensus assay.
We can then remove the NoCalls (no reads) from the individual vectors, further reducing the amount of memory needed.
}
