% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparatingMatrixToList.R
\name{SeparatingMatrixToList}
\alias{SeparatingMatrixToList}
\title{SeparatingMatrixToList}
\usage{
SeparatingMatrixToList(
  row_use,
  total_matrix,
  consensus,
  assay_to_split = "consensus",
  remove_nocalls = TRUE
)
}
\arguments{
\item{row_use}{The row the separate.}

\item{total_matrix}{The matrix to be split.}

\item{assay_to_split}{Which assay are we splitting?}

\item{remove_nocalls}{Do you want to remove NoCall cells?}
}
\description{
We separate a matrix of variant information to a list.
Each variant is an entry in the list.
NoCalls (cells with no reads covering a variant) can be removed.
This function gets called by RowWiseSplit in return.
}
