% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetVariantInfo.R
\name{SetVariantInfo}
\alias{SetVariantInfo}
\title{GetVariantInfo}
\usage{
SetVariantInfo(SE, seurat_object, information = "consensus", variants = NULL)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{seurat_object}{The Seurat object.}

\item{information}{The assay with the desired information. Default: consensus}

\item{variants}{A vector of variants.}
}
\description{
We add the genotyping information for a set of variants to a Seurat object.
The function returns a matrix with the values from the specified assay.
}
