% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetVariantInfo.R
\name{SetVariantInfo}
\alias{SetVariantInfo}
\title{SetVariantInfo}
\usage{
SetVariantInfo(
  SE,
  seurat_object,
  information = "consensus",
  variants = NULL,
  consensus_character = TRUE,
  consensus_group = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{seurat_object}{The Seurat object.}

\item{information}{The assay with the desired information. Default: consensus}

\item{variants}{A vector of variants.}

\item{consensus_character}{Should the consensus information be save as a character value?}

\item{consensus_group}{Should the consensus groups Alt and Both be merged? Has no effect if information is not consensus.}
}
\description{
We add the genotyping information for a set of variants to a Seurat object.
The function returns a Seurat object with the values from the specified assay added as meta data.
}
