% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantCloneSizeThresholding.R
\name{VariantCloneSizeThresholding}
\alias{VariantCloneSizeThresholding}
\title{VariantCloneSizeThresholding}
\usage{
VariantCloneSizeThresholding(
  SE,
  min_coverage = 2,
  fraction_negative_cells = 0.9,
  min_clone_size = 10,
  vaf_threshold = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{min_coverage}{Minimum coverage a variant needs to have.}

\item{fraction_negative_cells}{The fraction of negative cells needed.}

\item{min_clone_size}{minimum number of cells.}

\item{vaf_threshold}{Variant Allele Threshold. Cells above this threshold are considered mutated.}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
We get variants of interest using a clone size thresholding.
Source: https://github.com/petervangalen/MAESTER-2021
}
