% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantCorrelationHeatmap.R,
%   R/VariantFisherTestHeatmap.R
\name{VariantCorrelationHeatmap}
\alias{VariantCorrelationHeatmap}
\title{We generate a heatmap showing the correlation of somatic variants with the MT variants.}
\usage{
VariantCorrelationHeatmap(
  fisher_results,
  output_path = NULL,
  patient,
  min_alt_cells = 5,
  min_oddsratio = 1,
  width_use = 2000,
  height_use = 2000,
  padding_use = c(165, 165, 2, 2)
)

VariantCorrelationHeatmap(
  fisher_results,
  output_path = NULL,
  patient,
  min_alt_cells = 5,
  min_oddsratio = 1,
  width_use = 2000,
  height_use = 2000,
  padding_use = c(165, 165, 2, 2)
)
}
\arguments{
\item{fisher_results}{Data.frame with the correlation results.}

\item{output_path}{Path to the output folder.}

\item{patient}{The patient for this heatmap.}

\item{min_alt_cells}{Minimum number of mutated cells needed, otherwise an association will not be plotted.}

\item{min_oddsratio}{Minimum correlation needed.}

\item{width_use}{Width of the heatmap in px.}

\item{height_use}{Height of the heatmap in px.}

\item{padding_use}{Space around the heatmap in mm. If this is to low, the variant names might be cut off.}

\item{correlation_results}{Data.frame with the correlation results.}

\item{min_correlation}{Minimum correlation needed.}
}
\description{
We generate a heatmap showing the correlation of somatic variants with the MT variants.

We generate a heatmap showing the Fisher test of somatic variants with the MT variants.
}
