% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantCorrelationHeatmap.R
\name{VariantCorrelationHeatmap}
\alias{VariantCorrelationHeatmap}
\title{VariantCorrelationHeatmap}
\usage{
VariantCorrelationHeatmap(
  correlation_results,
  output_path = NULL,
  patient,
  min_alt_cells = 5,
  min_correlation = 0.5,
  width_use = 2000,
  height_use = 2000,
  padding_use = c(165, 165, 2, 2),
  verbose = TRUE
)
}
\arguments{
\item{correlation_results}{Data.frame with the correlation results.}

\item{output_path}{Path to the output folder.}

\item{patient}{The patient for this heatmap.}

\item{min_alt_cells}{Minimum number of mutated cells needed, otherwise a correlation will not be plotted.}

\item{min_correlation}{Minimum correlation needed.}

\item{width_use}{Width of the heatmap in px.}

\item{height_use}{Height of the heatmap in px.}

\item{padding_use}{Space around the heatmap in mm. If this is to low, the variant names might be cut off.}

\item{verbose}{Should the function be verbose? Default = TRUE}
}
\description{
We generate a heatmap showing the correlation of somatic variants with the MT variants.
Packages I want to remove. I cannot see where they are used.
ggplot2 parallel rcompanion tidyr
}
