% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantQuantileThresholding.R
\name{VariantQuantileThresholding}
\alias{VariantQuantileThresholding}
\title{We get variants of interest using the quantile thresholding.}
\usage{
VariantQuantileThresholding(
  SE,
  min_coverage = 2,
  quantiles = c(0.1, 0.9),
  thresholds = c(0.1, 0.9)
)
}
\arguments{
\item{SE}{SummarizedExperiment object.}

\item{min_coverage}{Minimum coverage needed.}

\item{quantiles}{The lower and upper quantile you want to use.}

\item{thresholds}{The VAF thresholds you want to use for the quantiles.}
}
\description{
We get variants of interest using the quantile thresholding.
}
