% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantQuantileThresholding.R
\name{VariantQuantileThresholding}
\alias{VariantQuantileThresholding}
\title{We get variants of interest using the quantile thresholding.}
\usage{
VariantQuantileThresholding(se, min_coverage = 2)
}
\arguments{
\item{se}{min_coverage}
}
\description{
We get variants of interest using the quantile thresholding.
}
