% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantWiseCorrelation.R
\name{VariantWiseCorrelation}
\alias{VariantWiseCorrelation}
\title{We get some descriptive statistics for each sample. We use the genotyping results for this.
fraction_interesting_subset.Rds
consensus_matrix_interesting_subset.Rds
Instead of running a loop, we run the script in a job array. We subset the matrices to only include the specific sample and
we calculate the correlations.}
\usage{
VariantWiseCorrelation(variants_list, n_cores = 1, p_value_adjustment = "fdr")
}
\arguments{
\item{variants_list}{n_cores p_value_adjustment}
}
\description{
We get some descriptive statistics for each sample. We use the genotyping results for this.
fraction_interesting_subset.Rds
consensus_matrix_interesting_subset.Rds
Instead of running a loop, we run the script in a job array. We subset the matrices to only include the specific sample and
we calculate the correlations.
}
